requirements.txt
pytest==7.2.2
selenium==4.8.2
ipython==8.11.0
python-dotenv==1.0.0
allure-pytest==2.13.1